Initial Commit - Working Heatmap
working renderHighHeatmap
working generic highchartsOutput with argument for type of highcharts
output to enable selective javascript file sourcing.

data and highcharts settings are returned in a list with named elements
corresponding to the highcharts settings to change.  To add more settings
simply add the settings to the highcharts javascript binding and set it's
value to input.<setting name>.